QuakeGuard/requirements.txt
streamlit>=1.32.0
langchain>=0.3.0
langchain-classic>=1.0.0
langchain-openai>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
openai>=1.40.0
pandas>=2.0.0
folium>=0.15.0
streamlit-folium>=0.18.0
requests>=2.31.0
python-dotenv>=1.0.0
pypdf>=4.0.0
streamlit-js-eval>=0.1.5
langchain-groq>=1.0.0
sentence-transformers>=3.0.0
